Fix CI results path; move stateful updateMaxAutoAssociationsWorks() spec out of async run section
