+++ .circleci/config.yml
@@ -1755,14 +1755,14 @@ jobs:
             set -x;
             set +e;
             cd /swirlds-platform/regression
-            CI_RESULTS_LOG=$(find results/4N_1C -name hapi-client-combined.log)
+            CI_RESULTS_LOG=$(find assets/results/4N_1C -name hapi-client-combined.log)
             grep $CI_RESULTS_LOG -e 'status=ERROR' -e 'status=FAILED'
             if [ "$?" -eq "0" ]; then
               exit 1
             else
               exit 0
             fi
-            SAMPLE_NODE_LOG=$(find results/4N_1C -name hgcaa.log | head -1)
+            SAMPLE_NODE_LOG=$(find assets/results/4N_1C -name hgcaa.log | head -1)
             grep $SAMPLE_NODE_LOG -e 'IssListener - In round'
             if [ "$?" -eq "0" ]; then
               exit 1
