+++ test-clients/src/main/java/com/hedera/services/bdd/suites/crypto/CryptoUpdateSuite.java
@@ -22,6 +22,7 @@ package com.hedera.services.bdd.suites.crypto;
 
 import com.hedera.services.bdd.spec.HapiApiSpec;
 import com.hedera.services.bdd.spec.HapiSpecSetup;
+import com.hedera.services.bdd.spec.assertions.ContractInfoAsserts;
 import com.hedera.services.bdd.spec.keys.KeyLabel;
 import com.hedera.services.bdd.spec.keys.KeyShape;
 import com.hedera.services.bdd.spec.keys.SigControl;
@@ -36,6 +37,7 @@ import java.util.concurrent.atomic.AtomicLong;
 
 import static com.hedera.services.bdd.spec.HapiApiSpec.defaultHapiSpec;
 import static com.hedera.services.bdd.spec.assertions.AccountDetailsAsserts.accountWith;
+import static com.hedera.services.bdd.spec.assertions.ContractInfoAsserts.contractWith;
 import static com.hedera.services.bdd.spec.keys.ControlForKey.forKey;
 import static com.hedera.services.bdd.spec.keys.KeyLabel.complex;
 import static com.hedera.services.bdd.spec.keys.KeyShape.SIMPLE;
@@ -45,11 +47,15 @@ import static com.hedera.services.bdd.spec.keys.SigControl.OFF;
 import static com.hedera.services.bdd.spec.keys.SigControl.ON;
 import static com.hedera.services.bdd.spec.queries.QueryVerbs.getAccountDetails;
 import static com.hedera.services.bdd.spec.queries.QueryVerbs.getAccountInfo;
+import static com.hedera.services.bdd.spec.queries.QueryVerbs.getContractInfo;
 import static com.hedera.services.bdd.spec.queries.QueryVerbs.getTxnRecord;
+import static com.hedera.services.bdd.spec.transactions.TxnVerbs.contractCreate;
+import static com.hedera.services.bdd.spec.transactions.TxnVerbs.contractUpdate;
 import static com.hedera.services.bdd.spec.transactions.TxnVerbs.cryptoCreate;
 import static com.hedera.services.bdd.spec.transactions.TxnVerbs.cryptoTransfer;
 import static com.hedera.services.bdd.spec.transactions.TxnVerbs.cryptoUpdate;
 import static com.hedera.services.bdd.spec.transactions.TxnVerbs.tokenCreate;
+import static com.hedera.services.bdd.spec.transactions.TxnVerbs.uploadInitCode;
 import static com.hedera.services.bdd.spec.transactions.token.TokenMovement.moving;
 import static com.hedera.services.bdd.spec.utilops.UtilVerbs.newKeyNamed;
 import static com.hedera.services.bdd.spec.utilops.UtilVerbs.overriding;
@@ -115,6 +121,7 @@ public class CryptoUpdateSuite extends HapiApiSuite {
 						updateFailsWithInvalidMaxAutoAssociations(),
 						usdFeeAsExpected(),
 						sysAccountKeyUpdateBySpecialWontNeedNewKeyTxnSign(),
+						updateMaxAutoAssociationsWorks()
 				}
 		);
 	}
@@ -432,6 +439,86 @@ public class CryptoUpdateSuite extends HapiApiSuite {
 								.hasPrecheck(BAD_ENCODING)
 				);
 	}
+
+	private HapiApiSpec updateMaxAutoAssociationsWorks() {
+		final int tokenAssociations_restrictedNetwork = 10;
+		final int tokenAssociations_adventurousNetwork = 1_000;
+		final int originalMax = 2;
+		final int newBadMax = originalMax - 1;
+		final int newGoodMax = originalMax + 1;
+		final String tokenA = "tokenA";
+		final String tokenB = "tokenB";
+
+		final String treasury = "treasury";
+		final String tokenACreate = "tokenACreate";
+		final String tokenBCreate = "tokenBCreate";
+		final String transferAToC = "transferAToC";
+		final String transferBToC = "transferBToC";
+		final String CONTRACT = "Multipurpose";
+		final String ADMIN_KEY = "adminKey";
+
+		return defaultHapiSpec("updateMaxAutoAssociationsWorks")
+				.given(
+						overridingTwo(
+								"entities.limitTokenAssociations", "true",
+								"tokens.maxPerAccount", "" + 10),
+						cryptoCreate(treasury)
+								.balance(ONE_HUNDRED_HBARS),
+						newKeyNamed(ADMIN_KEY),
+						uploadInitCode(CONTRACT),
+						contractCreate(CONTRACT)
+								.adminKey(ADMIN_KEY)
+								.maxAutomaticTokenAssociations(originalMax),
+						tokenCreate(tokenA)
+								.tokenType(TokenType.FUNGIBLE_COMMON)
+								.initialSupply(Long.MAX_VALUE)
+								.treasury(treasury)
+								.via(tokenACreate),
+						getTxnRecord(tokenACreate)
+								.hasNewTokenAssociation(tokenA, treasury),
+						tokenCreate(tokenB)
+								.tokenType(TokenType.FUNGIBLE_COMMON)
+								.initialSupply(Long.MAX_VALUE)
+								.treasury(treasury)
+								.via(tokenBCreate),
+						getTxnRecord(tokenBCreate)
+								.hasNewTokenAssociation(tokenB, treasury),
+						getContractInfo(CONTRACT)
+								.has(ContractInfoAsserts.contractWith().maxAutoAssociations(originalMax))
+								.logged()
+				)
+				.when(
+						cryptoTransfer(moving(1, tokenA).between(treasury, CONTRACT))
+								.via(transferAToC),
+						getTxnRecord(transferAToC)
+								.hasNewTokenAssociation(tokenA, CONTRACT),
+						cryptoTransfer(moving(1, tokenB).between(treasury, CONTRACT))
+								.via(transferBToC),
+						getTxnRecord(transferBToC)
+								.hasNewTokenAssociation(tokenB, CONTRACT)
+				)
+				.then(
+						getContractInfo(CONTRACT)
+								.payingWith(GENESIS)
+								.has(contractWith()
+										.hasAlreadyUsedAutomaticAssociations(originalMax)
+										.maxAutoAssociations(originalMax)),
+						contractUpdate(CONTRACT)
+								.newMaxAutomaticAssociations(newBadMax)
+								.hasKnownStatus(EXISTING_AUTOMATIC_ASSOCIATIONS_EXCEED_GIVEN_LIMIT),
+						contractUpdate(CONTRACT)
+								.newMaxAutomaticAssociations(newGoodMax),
+						contractUpdate(CONTRACT)
+								.newMaxAutomaticAssociations(tokenAssociations_restrictedNetwork + 1)
+								.hasKnownStatus(REQUESTED_NUM_AUTOMATIC_ASSOCIATIONS_EXCEEDS_ASSOCIATION_LIMIT),
+
+						overriding("entities.limitTokenAssociations", "false"),
+						contractUpdate(CONTRACT)
+								.newMaxAutomaticAssociations(tokenAssociations_restrictedNetwork + 1),
+						overriding("tokens.maxPerAccount", "" + tokenAssociations_adventurousNetwork)
+				);
+	}
+
 
 	@Override
 	protected Logger getResultsLogger() {
